requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
gunicorn==23.0.0
markdown2==2.5.3
python-frontmatter==1.1.0
jinja2==3.1.6
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
fastapi-mail==1.5.0
